Correction d'un bug qui fesait crash le serveur. Gestion du cache.

+++ controllers/announcementcontroller.js
@@ -28,6 +28,7 @@ exports.getList = async(req, res) => {
         const decoded = jwt.verify(token, config.secret)
 
         const announcements = await Announcement.find( { author: decoded.id } );
+        res.setHeader("Cache-Control", "no-cache")
         res.status(200).json(announcements)
     } catch (err) {
         res.status(500).send("Impossible de récupérer la liste des annnonces.")
@@ -39,16 +40,18 @@ exports.getDetails = async(req, res) => {
 
         announcement = await checkAuthorization(req, res);
 
-        const announcementJson = JSON.stringify(announcement);
-        const hash = etag(announcementJson);
-        if (req.headers['if-none-match'] === hash) {
-            return res.status(304).send("L'annonce n' pas été modifié");
+        if (announcement instanceof Announcement){
+            const announcementJson = JSON.stringify(announcement);
+            const hash = etag(announcementJson);
+            if (req.headers['if-none-match'] === hash) {
+                return res.status(304).send("L'annonce n' pas été modifié");
+            }
+            res.setHeader('ETag', hash);
+            res.status(200).json(announcement);
         }
-        res.setHeader('ETag', hash);
-        res.status(200).json(announcement);
     } catch (err) {
         console.log(err);
-        res.status(500).send("Impossible de récupérer les détails de l'annonce getDetails")
+        res.status(500).send("Impossible de récupérer les détails de l'annonce getDetails");
     }
 }
 
@@ -58,20 +61,22 @@ exports.update = async(req, res) => {
 
         announcement = await checkAuthorization(req, res);
 
-        const clientETag = req.headers['if-match'];
-        const currentETag = etag(JSON.stringify(announcement));
-        console.log(clientETag);
-        console.log(currentETag);
-        if (clientETag !== currentETag) {
-            return res.status(412).send("Precondition Failed: L'Etag ne correspond pas."); // 412 Precondition Failed
+        if (announcement instanceof Announcement){
+            const clientETag = req.headers['if-match'];
+            const currentETag = etag(JSON.stringify(announcement));
+            console.log(clientETag);
+            console.log(currentETag);
+            if (clientETag !== currentETag) {
+                return res.status(412).send("Precondition Failed: L'Etag ne correspond pas."); // 412 Precondition Failed
+            }
+
+            announcement.title = req.body.title || announcement.title;
+            announcement.description = req.body.description || announcement.description;
+            announcement.image = req.body.image || announcement.image;
+
+            const updateAnnouncement = await announcement.save();
+            res.status(200).json(updateAnnouncement);
         }
-
-        announcement.title = req.body.title || announcement.title;
-        announcement.description = req.body.description || announcement.description;
-        announcement.image = req.body.image || announcement.image;
-
-        const updateAnnouncement = await announcement.save();
-        res.status(200).json(updateAnnouncement);
     } catch(err){
         console.log(err);
         res.status(500).send("Une erreur est survenue lors de la modification de l'annonce.")
@@ -84,8 +89,10 @@ exports.delete = async(req, res) => {
 
         announcement = await checkAuthorization(req, res);
 
-        await announcement.deleteOne({ id: announcement._id });
-        res.status(200).send("Annonce supprimée avec succés.");
+        if (announcement instanceof Announcement){
+            await announcement.deleteOne({ id: announcement._id });
+            res.status(200).send("Annonce supprimée avec succés.");
+        }
     } catch(err) {
         console.log(err);
         res.status(500).send("Une erreur est survenue lors de la suppression de l'annonce.")
@@ -98,9 +105,11 @@ async function checkAuthorization(req, res){
 
     const announcement = await Announcement.findById(req.params.id);
     if(!announcement){
-        return res.status(404).send("Impossible de trouver l'annonce.");
+        res.status(404).send("Impossible de trouver l'annonce.");
+        return false;
     } else if(announcement.author != decoded.id){
-        return res.status(403).send("Vous n'avez pas les permissions pour effectuer cette action.")
+        res.status(403).send("Vous n'avez pas les permissions pour effectuer cette action.");
+        return false;
     }
     return announcement;
 }
